Check p2p address as valid anemo address
