+++ crates/sui-types/src/sui_system_state/mod.rs
@@ -14,6 +14,7 @@ use move_core_types::language_storage::TypeTag;
 use move_core_types::value::MoveTypeLayout;
 use move_core_types::{ident_str, identifier::IdentStr, language_storage::StructTag};
 use move_vm_types::values::Value;
+use multiaddr::Multiaddr;
 use serde::{Deserialize, Serialize};
 use std::collections::BTreeMap;
 use tracing::error;
@@ -189,4 +190,25 @@ where
 
 pub fn get_sui_system_state_version(_protocol_version: ProtocolVersion) -> u64 {
     INIT_SYSTEM_STATE_VERSION
+}
+
+pub fn multiaddr_to_anemo_address(multiaddr: &Multiaddr) -> Option<anemo::types::Address> {
+    use multiaddr::Protocol;
+    let mut iter = multiaddr.iter();
+
+    match (iter.next(), iter.next(), iter.next()) {
+        (Some(Protocol::Ip4(ipaddr)), Some(Protocol::Udp(port)), None) => {
+            Some((ipaddr, port).into())
+        }
+        (Some(Protocol::Ip6(ipaddr)), Some(Protocol::Udp(port)), None) => {
+            Some((ipaddr, port).into())
+        }
+        (Some(Protocol::Dns(hostname)), Some(Protocol::Udp(port)), None) => {
+            Some((hostname.as_ref(), port).into())
+        }
+        _ => {
+            tracing::debug!("unsupported p2p multiaddr: '{multiaddr}'");
+            None
+        }
+    }
 }
