+++ crates/sui-types/src/sui_system_state/sui_system_state_inner_v1.rs
@@ -109,6 +109,10 @@ impl ValidatorMetadataV1 {
             .map_err(|_| E_METADATA_INVALID_NET_ADDR)?;
         let p2p_address = Multiaddr::try_from(self.p2p_address.clone())
             .map_err(|_| E_METADATA_INVALID_P2P_ADDR)?;
+        // Also make sure that the p2p address is a valid anemo address.
+        // TODO: This will trigger a bunch of Move test failures today since we did not give proper
+        // value for p2p address.
+        // multiaddr_to_anemo_address(&p2p_address).ok_or(E_METADATA_INVALID_P2P_ADDR)?;
         let primary_address = Multiaddr::try_from(self.primary_address.clone())
             .map_err(|_| E_METADATA_INVALID_PRIMARY_ADDR)?;
         let worker_address = Multiaddr::try_from(self.worker_address.clone())
